pyproject.toml
[project]
name = "maxai-mcp"
description = "MCP server that connects to Max AI agents"
readme = "README.md"
requires-python = ">=3.10.7"
dynamic = ["version"]
dependencies = [
    "answerrocket-client==0.2.78",
    "mcp>=1.25.0",
    "starlette>=0.49.1",
    "starlette-context==0.4.0",
]

[build-system]
requires = ["pip>=25.3", "setuptools>=78.1.1"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["lib*", "scripts*"]

[tool.setuptools.dynamic]
version = { attr = "mcp_server.__version__" }

[project.scripts]
maxai-mcp = "mcp_server.__main__:main"
